# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Government Programs API",
    "version": "1.0.0",
    "description": "Phase 1 public API for discovering government incentive programs."
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Production (example placeholder)"
    }
  ],
  "paths": {
...
    "/v1/programs": {
      "get": {
        "summary": "List programs",
        "operationId": "listPrograms",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Full-text search query"
          },
          {
            "name": "country",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "US",
                "CA"
              ]
            }
          },
          {
            "name": "state_province",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "State or province code (e.g., WA)"
          },
          {
            "name": "industry[]",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "benefit_type[]",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "status[]",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "updated_at",
                "-updated_at",
                "title",
                "-title"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "rank",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ltr"
              ]
            },
            "description": "Optional learning-to-rank strategy"
          },
          {
            "name": "rank_n",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Number of candidates to rerank before paging"
          }
        ],
        "responses": {
          "200": {
            "description": "Programs response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/programs/{id}": {
      "get": {
        "summary": "Get a program by ID or UID",
        "operationId": "getProgram",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Program UID or numeric identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Program detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Program not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProgramResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uid": {
            "type": "string"
          },
          "country_code": {
            "type": "string",
            "enum": [
              "US",
              "CA"
            ]
          },
          "authority_level": {
            "type": "string",
            "enum": [
              "federal",
              "state",
              "prov",
              "territory",
              "regional",
              "municipal"
            ]
          },
          "jurisdiction_code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "benefit_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiBenefitType"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiProgramStatus"
          },
          "industry_codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_id": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "number"
          },
          "updated_at": {
            "type": "number"
          },
          "benefits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramBenefitResponse"
            }
          },
          "criteria": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramCriterionResponse"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "uid",
          "country_code",
          "authority_level",
          "jurisdiction_code",
          "title",
          "status",
          "industry_codes",
          "created_at",
          "updated_at",
          "benefits",
          "criteria",
          "tags"
        ],
        "additionalProperties": false
      },
      "ApiBenefitType": {
        "type": "string",
        "enum": [
          "grant",
          "rebate",
          "tax_credit",
          "loan",
          "guarantee",
          "voucher",
          "other"
        ]
      },
...
      "ProgramListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramResponse"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": {
                "type": "number"
              },
              "page": {
                "type": "number"
              },
              "pageSize": {
                "type": "number"
              },
              "ranking": {
                "type": "string"
              }
            },
            "required": [
              "total",
              "page",
              "pageSize"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
      "ProgramDetailResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "number"
          },
          "uid": {
            "type": "string"
          },
          "country_code": {
            "type": "string",
            "enum": [
              "US",
              "CA"
            ]
          },
          "authority_level": {
            "type": "string",
            "enum": [
              "federal",
              "state",
              "prov",
              "territory",
              "regional",
              "municipal"
            ]
          },
          "jurisdiction_code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "benefit_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiBenefitType"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiProgramStatus"
          },
          "industry_codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_id": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "number"
          },
          "updated_at": {
            "type": "number"
          },
          "benefits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramBenefitResponse"
            }
          },
          "criteria": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgramCriterionResponse"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "authority_level",
          "benefits",
          "country_code",
          "created_at",
          "criteria",
          "id",
          "industry_codes",
          "jurisdiction_code",
          "status",
          "tags",
          "title",
          "uid",
          "updated_at"
        ]
      },
...
    }
  }
}